requirements.txt
fastapi==0.115.5
uvicorn==0.31.1
sqlmodel==0.0.22
pyyaml==6.0.2
jinja2==3.1.4
python-multipart==0.0.20
